requirements.txt
appscript==1.0.1
cycler==0.10.0
matplotlib==2.1.1
mazelib==0.7
numpy==1.14.0rc1
opencv-python==3.3.0.10
Pillow==4.3.0
PyAutoGUI==0.9.36
pygame==1.9.3
PyMsgBox==1.0.6
pynput==1.3.8.1
pyobjc==4.1
pyparsing==2.2.0
PyScreeze==0.1.13
python-dateutil==2.6.1
PyTweening==1.0.3
pytz==2017.3
selenium==3.8.0